[LCR-280] refactor: dto 패키지 이동
