+++ src/main/java/com/lgcns/domain/survey/kafka/consumer/MemberAnswerConsumer.java
@@ -3,8 +3,8 @@ package com.lgcns.domain.survey.kafka.consumer;
 import com.lgcns.domain.survey.domain.MemberAnswer;
 import com.lgcns.domain.survey.domain.Survey;
 import com.lgcns.domain.survey.exception.SurveyErrorCode;
-import com.lgcns.domain.survey.kafka.dto.SurveyChoiceDto;
 import com.lgcns.domain.survey.kafka.message.MemberAnswerMessage;
+import com.lgcns.domain.survey.kafka.message.dto.SurveyChoiceDto;
 import com.lgcns.domain.survey.repository.MemberAnswerRepository;
 import com.lgcns.domain.survey.repository.SurveyRepository;
 import com.lgcns.global.error.exception.CustomException;
