+++ src/main/java/com/lgcns/domain/survey/kafka/message/dto/SurveyChoiceDto.java
@@ -1,4 +1,4 @@
-package com.lgcns.domain.survey.kafka.dto;
+package com.lgcns.domain.survey.kafka.message.dto;
 
 public record SurveyChoiceDto(Long surveyId, Long choiceId) {
     public static SurveyChoiceDto of(Long surveyId, Long choiceId) {
