+++ src/test/java/com/lgcns/domain/survey/kafka/MemberAnswerConsumerTest.java
@@ -11,8 +11,8 @@ import com.lgcns.domain.popup.domain.Popup;
 import com.lgcns.domain.popup.repository.PopupRepository;
 import com.lgcns.domain.survey.domain.MemberAnswer;
 import com.lgcns.domain.survey.domain.Survey;
-import com.lgcns.domain.survey.kafka.dto.SurveyChoiceDto;
 import com.lgcns.domain.survey.kafka.message.MemberAnswerMessage;
+import com.lgcns.domain.survey.kafka.message.dto.SurveyChoiceDto;
 import com.lgcns.domain.survey.repository.MemberAnswerRepository;
 import com.lgcns.domain.survey.repository.SurveyRepository;
 import java.time.Duration;
